Separate state logic to a hook
